pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.0.1"
description = "flask project page analyzer"
authors = ["sosnin valentin"]
license = "MIT"
readme = "README.md"
packages = [
    {include = "page_analyzer"},
]

[tool.poetry.dependencies]
python = "^3.8.10"
flask = "^3.0.2"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
jinja2 = "^3.1.3"
psycopg2-binary = "^2.9.9"
validators = "^0.28.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
